Extract PlantNectar into its own module
PlantNectar is a Food, not a Plant — it only lived in plant.py for
historical reasons, padding that module to 879 lines. Move it to
core/entities/plant_nectar.py so plant.py is solely about the Plant
entity (now 728 lines; re-pin the god-class ratchet accordingly).

plant.py keeps importing PlantNectar (it constructs nectar in
_try_produce_nectar), which doubles as the backwards-compat re-export so
existing `from core.entities.plant import PlantNectar` callers are
unaffected. The new module only needs Plant for a type hint, so the
dependency is one-directional at runtime (no import cycle).

Point core.entities.__init__ at the canonical module, allowlist the new
module in the energy-mutation guard (PlantNectar writes self.energy
directly, like other Food), and extend the infra import-guard so `from
core.entities.plant_nectar import PlantNectar` can't sneak
concrete-entity coupling back into backend/shared code.
