+++ core/entities/__init__.py
@@ -3,7 +3,8 @@
 from core.entities.base import Agent, Castle, Entity, LifeStage, Rect
 from core.entities.fish import Fish
 from core.entities.generic_agent import AgentComponents, GenericAgent
-from core.entities.plant import Plant, PlantNectar
+from core.entities.plant import Plant
+from core.entities.plant_nectar import PlantNectar
 from core.entities.predators import Crab
 from core.entities.resources import Food, LiveFood
